Ditched the OAuth Dynamic Client Registration spec in favour of a custom approach, as it really doesn't meet our needs.
The dynamic client registration spec is pretty clearly designed for
OAuth clients which are used by multiple users (i.e. a service
registers dynamically as a client of another service). I thought we
could make it fit our use cases, but it doesn't allow things like (a)
renaming a device from another device, or (b) deleting a device from
another device, etc. It's also overly complex and we wouldn't be
unconditionally compliant with the RFC anyway due to requiring
authentication to register a client. Need to amend our auth spec to
reflect this.