Ensured responses aren't cacheable
None of the auth server responses should be cached, so added tests to
ensure that. Added the cucumber-rest and http_capture libraries to help
out with this, and switched from Mechanize to HTTParty as the testing
http client to make things cleaner.

Also ensured that the Date header is set as some Rack middleware doesn't
set it, and it's required to be returned by the HTTP spec.