Can now get client info for user
The main change is to be able to list clients for a user and get info
about a specific client - however, I noticed some inconsistencies with
the way I'd done user information while doing this so there are some
changed related to that, which make them consistent.

Note that a couple of the test steps for negative test cases are
incomplete or undefined at the moment because the tests were originally
written with the concept of "the current user" and "the current device"
which doesn't support some of the more complex negative test cases
needed like "another user's device" so I will need to do a fair bit of
restructuring of how the tests store the current state to permit this
(i.e. in lists). I'm going to do that as a separate change though.