Removed access_tokens DB table
Rather than tracking all the access token identifiers, simply got rid of
the table and added the refresh token identifier into the token claims.
There's no point in having an access token identifier in there as the
token isn't single-use so it can't be used to defend against replay
attacks, and given we trust the claims it makes sense just to put an
'internal use' claim showing the source of the token (i.e. the issuing
refresh token on which it is based).