Fixed test semantics around token usage
A number of the tests were using "my current client" implying that the
tests were dealing with access tokens and refresh tokens that were bound
to the client - however, in many cases the tokens had not been bound.
This doesn't actually affect the tests done so far (as the ones where it
was important already did it explicitly) but now the tests do what they
say. It does become important for things like client deregistration,
hence why it was noticed now.