Fixed client_secret only not returning 400
When only client_id or client_secret are returned, then the auth server
should return a 400 Bad Request indicating that both of them are
needed. At the moment this works for client_id but not for
client_secret. This is the fix, plus some more tests for the missed
scenarios.