Changed table structure to have no PK
The login_attempt table now has no PK but has a covering index on the
fields we’re interested in. Also changed the error handling approach to
the one suggested by JP which seems a lot neater. Note that at the
moment there is no proper logging which is due to Sinatra/Rack logging
being a bit ‘special’ so will do that as a separate ticket.